requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
alembic==1.6.5
click==7.1.2
flask-cors==3.0.8
flask-login==0.5.0
flask-migrate==3.0.1
flask-sqlalchemy==2.5.1
flask-wtf==0.15.1
flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
jinja2==3.0.1
mako==1.1.4
markupsafe==2.0.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
sqlalchemy==1.4.19
werkzeug==2.0.1
wtforms==2.3.3
openai==0.11.0